Typescript implementation - directory-item & grid-section components completed
